Create 3a
Create, test, and validate an XHTML document that defines a table that
has two levels of column labels: an overall label, Meals, and three
secondary labels, Breakfast, Lunch, and Dinner. There must be two levels
of row labels: an overall label, Foods, and four secondary labels, Bread,
Main Course, Vegetable, and Dessert. The cells of the table must contain
a number of grams for each of the food categories